extract constant to avoid circular import
