+++ app/constants.py
@@ -1,0 +1,1 @@
+MAX_VALOR_NEGOCIADO = 9_999_999
